requirements.txt
aeon==1.3.0
baycomp==1.0.3
matplotlib==3.10.8
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.7.2
seaborn==0.13.2
torch==2.9.1
tsfresh==0.21.1
umap-learn==0.5.11
